+++ R/param.R
@@ -37,7 +37,7 @@ get_param_files <- function(param_files) {
 #' scenarios <- get_param_dirs(param_files)
 #' scenarios
 #' scenario <- scenarios[1]
-#' run_engine_contin(scenario, n_sims = 2, verbose = FALSE)
+#' run_engine_contin(scenario, n_sims = 2, verbose = FALSE, version = "6.2.5")
 #' read_patients(scenario)
 #' }
 get_param_dirs <- function(param_files) {