+++ tests/testthat/test-run_engine_contin.R
@@ -2,7 +2,7 @@ test_that("run_engine_contin() with default args", {
   skip_paths()
   facts_file <- get_facts_file_example("contin.facts")
   out <- run_flfll(facts_file, verbose = FALSE)
-  run_engine_contin(out, n_sims = 1L, verbose = FALSE)
+  run_engine_contin(out, n_sims = 1L, verbose = FALSE, version = "6.2.5")
   files <- list.files(out, pattern = "patients[0-9]+\\.csv$", recursive = TRUE)
   expect_equal(length(files), 4L)
 })
@@ -12,11 +12,21 @@ test_that("run_engine_contin() on actual param files", {
   facts_file <- get_facts_file_example("contin.facts")
   out <- run_flfll(facts_file, verbose = FALSE)
   param_files <- get_param_files(out)
-  run_engine_contin(param_files[1], n_sims = 1L, verbose = FALSE)
+  run_engine_contin(
+    param_files[1],
+    n_sims = 1L,
+    verbose = FALSE,
+    version = "6.2.5"
+  )
   pattern <- "patients[0-9]+\\.csv$"
   files <- list.files(out, pattern = pattern, recursive = TRUE)
   expect_equal(length(files), 1L)
-  run_engine_contin(param_files, n_sims = 1L, verbose = FALSE)
+  run_engine_contin(
+    param_files,
+    n_sims = 1L,
+    verbose = FALSE,
+    version = "6.2.5"
+  )
   files <- list.files(out, pattern = pattern, recursive = TRUE)
   expect_equal(length(files), 4L)
 })
@@ -31,7 +41,12 @@ test_that("run_engine_contin() on actual param files in param file dir", {
   skip_if_not_installed("withr")
   withr::with_dir(
     param_dir,
-    run_engine_contin(param_file, n_sims = 1L, verbose = FALSE)
+    run_engine_contin(
+      param_file,
+      n_sims = 1L,
+      verbose = FALSE,
+      version = "6.2.5"
+    )
   )
   pattern <- "patients[0-9]+\\.csv$"
   files <- list.files(out, pattern = pattern, recursive = TRUE)
@@ -54,7 +69,7 @@ test_that("run_engine_contin() with lots of sims", {
     flfll_seed = 1,
     verbose = FALSE
   )
-  run_engine_contin(out, n_sims = 1e3, verbose = FALSE)
+  run_engine_contin(out, n_sims = 1e3, verbose = FALSE, version = "6.2.5")
   files <- list.files(out, pattern = "weeks[0-9]+\\.csv$", recursive = TRUE)
   expect_equal(length(files), 4 * n_weeks)
   pattern <- "patients[0-9]+\\.csv$"
@@ -74,14 +89,24 @@ test_that("run_engine_contin() with no param files", {
     tmp, {
       file.create("abc")
       expect_error(
-        run_engine_contin(param_files = "abc", n_sims = 1L, verbose = FALSE),
+        run_engine_contin(
+          param_files = "abc",
+          n_sims = 1L,
+          verbose = FALSE,
+          version = "6.2.5"
+        ),
         regexp = msg
       )
       dir.create("x")
       file.create(file.path("x", "y"))
       for (f in c("x", file.path("x", "y"))) {
         expect_error(
-          run_engine_contin(param_files = f, n_sims = 1L, verbose = FALSE),
+          run_engine_contin(
+            param_files = f,
+            n_sims = 1L,
+            verbose = FALSE,
+            version = "6.2.5"
+          ),
           regexp = msg
         )
       }